ok now what

+++ src/PlayerObject.cpp
@@ -9,7 +9,7 @@ class $modify(MyPlayerObject, PlayerObject) {
 		UniversalJumpscareSprite* unjus = Utils::getUNJUS();
 
 		if (Utils::modEnabled() && this->isPlayer1() && button == PlayerButton::Jump && this->m_gameLayer && !this->m_gameLayer->m_isEditor && unjus->isVisible() && !unjus->shouldStop() && !Manager::shouldNotJumpscare(JumpscareType::GameplayClick)) {
-			unjus->play(JumpscareType::GameplayClick); // the enum is there to prevent timer from being unintentionally reset
+			unjus->play(static_cast<int>(JumpscareType::GameplayClick)); // the enum is there to prevent timer from being unintentionally reset
 		}
 
 		return ret;
@@ -19,7 +19,7 @@ class $modify(MyPlayerObject, PlayerObject) {
 		UniversalJumpscareSprite* unjus = Utils::getUNJUS();
 
 		if (Utils::modEnabled() && this->isPlayer1() && this->m_isDead && this->m_gameLayer && !this->m_gameLayer->m_isEditor && unjus->isVisible() && !unjus->shouldStop() && !Manager::shouldNotJumpscare(JumpscareType::GameplayDeath)) {
-			unjus->play(JumpscareType::GameplayDeath); // the enum is there to prevent timer from being unintentionally reset
+			unjus->play(static_cast<int>(JumpscareType::GameplayDeath)); // the enum is there to prevent timer from being unintentionally reset
 		}
 
 		return ret;
